+++ src/ENDFtk/section/33/test/33.test.cpp
@@ -36,7 +36,7 @@ SCENARIO( "section::Type< 33 >" ) {
       double lower = 2.500000e+3;
       double upper = 2.000000e+7;
       std::vector< double > coefficients = {1, -1, -1, -1, -1, -1, -1};
-      std::vector< double > reactions = {1, 4, 16, 17, 18, 37, 102};
+      std::vector< int > reactions = {1, 4, 16, 17, 18, 37, 102};
       std::vector< DerivedCovariance > nc = {
         DerivedRedundant( lower, upper,
                           std::move(coefficients),
